requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
ruamel.yaml==0.18.6
requests==2.32.3
scikit-learn==1.5.2
numpy==2.1.1
prophet==1.1.5 ; sys_platform != 'win32'  # optional; skip on Windows dev by default

